Restructure controllers package and tests
 * Move tests/testUtils/test_util.go to controllers/controllers_test.go
   as this dependency was outrageous.
 * Set up handlers were there implementation lives. The controllers
   package now exposes a single gorilla/mux.Router where all routes are
   registered.
 * Remove horrible globals that were not guarded (like lastError and
   lastStatus).
 * Use a single aetest.Instance for all testing (I think this is needed
   to have a consistent Datastore; not sure).
