+++ controllers/result.go
@@ -17,6 +17,11 @@ import (
 	"golang.org/x/net/context"
 )
 
+func init() {
+	router.HandleFunc("/results", setup(CreateResult))
+	router.HandleFunc("/results/{resultKey}", setup(GetResult))
+}
+
 // CreateResult saves a new result when a coder starts a challenge.
 func CreateResult(ctx context.Context, w http.ResponseWriter, r *http.Request) (int, error) {
 	if r.Method != "POST" {
