+++ controllers/status.go
@@ -41,6 +41,7 @@ var initTime time.Time
 
 func init() {
 	initTime = time.Now()
+	router.HandleFunc("/status", hsts(Status))
 }
 
 // Status gathers a quick overview of the system state
