+++ controllers/task.go
@@ -13,6 +13,12 @@ import (
 	"google.golang.org/appengine/datastore"
 )
 
+func init() {
+	router.HandleFunc("/tasks/{key}", setup(TaskByKey))
+	router.HandleFunc("/tasks/{key}/tests", setup(TestsByTaskKey))
+	router.HandleFunc("/tasks", setup(Tasks))
+}
+
 // TaskByKey loads a task by key.
 func TaskByKey(ctx context.Context, w http.ResponseWriter, r *http.Request) (status int, err error) {
 	p, ok := passenger.FromContext(ctx)
