--- util/routing/routing.go
@@ -1,13 +1,0 @@
-package routing
-
-import (
-	"net/http"
-
-	"golang.org/x/net/context"
-)
-
-// ContextHandlerFunc is similar to a http.HandlerFunc, but also gets passed
-// the current context.
-// To ease error handling, a ContextHandleFunc must return a HTTP status
-// code and an error. Still, the handler is allowed to write a response.
-type ContextHandlerFunc func(context.Context, http.ResponseWriter, *http.Request) (int, error)
